Implementación de UserUtil
Implementaciones de la clase UserUtil que permite acceder tanto al User autenticado desde cualquier clase como saber si el User autenticado es Admin o Cliente
